refactor. gi TallIInput korrekt navn TallInput
Retter opp feilstavet komponentnavn fra TallIInput til TallInput, og oppdaterer importstien i AntallDagerInput.
