+++ src/komponenter/form/TallInput.tsx
